+++ packages/flutter_chat_core/lib/flutter_chat_core.dart
@@ -11,3 +11,4 @@ export 'src/theme/chat_theme.dart';
 export 'src/theme/chat_theme_extension.dart';
 export 'src/utils/is_only_emoji.dart';
 export 'src/utils/typedefs.dart';
+export 'src/utils/user_cache.dart';