+++ packages/flutter_chat_core/lib/src/utils/typedefs.dart
@@ -1,6 +1,9 @@
 import 'package:flutter/widgets.dart';
 
 import '../models/message.dart';
+import '../models/user.dart';
+
+typedef ResolveUserCallback = Future<User?> Function(String id);
 
 typedef ChatItem =
     Widget Function(