+++ packages/flutter_chat_ui/lib/src/utils/typedefs.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_chat_core/flutter_chat_core.dart';
 
-typedef ResolveUserCallback = Future<User?> Function(String id);
 typedef OnMessageTapCallback =
     void Function(Message message, {int index, TapUpDetails details});
 typedef OnMessageSendCallback = void Function(String text);